Clean up
1.  Remove proptypes from CallWindowPeer
2. ControlRoom Window
3. AddPeerToCascade
4. Refactored state into own file
