+++ client/src/js/app/index.js
@@ -5,41 +5,12 @@ import { toast } from 'react-toastify';
 import labeledStream from '../streamutils/labeledStream'
 import actions from './actions'
 import effects from './effects'
+import state from './state'
 import { proxyMethods, setProxyActions } from './proxyMethods'
 export { proxyMethods }
 
 logLoader(module);
-const state = {
-    title: 'This title',
-    diags: [],
-    showCascade: false,
-    showControlRoom: false,
-    members: [], //array of member session numbers
-    cascade: [], //array of session numbers in order
-    callInfo: {},
-    sessions: {},
-    allSessions: [],
-    users: {}, // keyed list of users with their data
-    _message: {
-        text: '',
-        delay: 1000
-    },
-    streams: {
-        empty: null,
-        local: null,
-        toControl: null,
-        cascade: null,
 
-    },
-    attrs: {},
-    events: [],
-    lastEvent: {},
-    control: null,
-    leader: null,
-    otherRoles: {
-
-    }
-}
 
 let theActions
 // console.log('conform source code', effects.socket.onConfirm + '')
